Optimize some parts of the code
