+++ rosbag2_cpp/src/rosbag2_cpp/service_utils.cpp
@@ -28,10 +28,16 @@ const size_t kServiceEventTypePostfixLen = strlen(kServiceEventTopicTypePostfix)
 
 bool is_service_event_topic(const std::string & topic_name, const std::string & topic_type)
 {
-  if ((topic_name.length() <= kServiceEventTopicPostfixLen) ||
-    (topic_name.substr(topic_name.length() - kServiceEventTopicPostfixLen) !=
-    RCL_SERVICE_INTROSPECTION_TOPIC_POSTFIX) ||
-    (topic_name.find("/_action/") != std::string::npos))
+  if (topic_name.find("/_action/") != std::string::npos) {
+    return false;
+  }
+
+  if (topic_name.length() <= kServiceEventTopicPostfixLen) {
+    return false;
+  }
+
+  if (topic_name.substr(topic_name.length() - kServiceEventTopicPostfixLen) !=
+    RCL_SERVICE_INTROSPECTION_TOPIC_POSTFIX)
   {
     return false;
   }
