+++ rosbag2_py/CMakeLists.txt
@@ -94,6 +94,7 @@ pybind11_add_module(_info
   src/rosbag2_py/format_bag_metadata.cpp
   src/rosbag2_py/format_action_info.cpp
   src/rosbag2_py/format_service_info.cpp
+  src/rosbag2_py/format_utils.cpp
   src/rosbag2_py/info_sorting_method.cpp
 )
 target_link_libraries(_info PUBLIC
