+++ rosbag2_py/src/rosbag2_py/format_utils.cpp
@@ -1,0 +1,40 @@
+// Copyright 2025 Sony Group Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <cstdint>
+#include <sstream>
+#include <string>
+#include <iomanip>
+
+#include "format_utils.hpp"
+
+namespace rosbag2_py
+{
+std::string format_file_size(uint64_t file_size)
+{
+  double size = static_cast<double>(file_size);
+  static const char * units[] = {"B", "KiB", "MiB", "GiB", "TiB"};
+  double reference_number_bytes = 1024;
+  int index = 0;
+  while (size >= reference_number_bytes && index < 4) {
+    size /= reference_number_bytes;
+    index++;
+  }
+
+  std::stringstream rounded_size;
+  int size_format_precision = index == 0 ? 0 : 1;
+  rounded_size << std::setprecision(size_format_precision) << std::fixed << size;
+  return rounded_size.str() + " " + units[index];
+}
+}  //  namespace rosbag2_py
