+++ rosbag2_py/src/rosbag2_py/format_utils.hpp
@@ -1,0 +1,26 @@
+// Copyright 2025 Sony Group Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef ROSBAG2_PY__FORMAT_UTILS_HPP_
+#define ROSBAG2_PY__FORMAT_UTILS_HPP_
+
+#include <cstdint>
+#include <string>
+
+namespace rosbag2_py
+{
+std::string format_file_size(uint64_t file_size);
+}  // namespace rosbag2_py
+
+#endif  // ROSBAG2_PY__FORMAT_UTILS_HPP_
